Refactor: move learn method to vw_base (and make it an abstract base class)
